All unpack() methods return a length field
All unpack() methods return an explicit length field, to facilitate parsing
of variable-sized records.  In individual cases (e.g., labels), this is not
strictly necessary since the length can be inferred from the resulting object.
But for other objects (e.g., Questions), the logic is more complicated.  So
for simplicity and consistency, all unpack() methods now return an explicit
length, so that the caller can easily resume or continue parsing the incoming
byte-stream regardless of the structures involved.